Introduce validation and exception handling for transaction requests
- Add `TransactionValidation` interface and `TransactionValidationImpl` to validate transaction requests.
- Add `ConstraintViolationException` and `ErrorType` enum to manage error codes and messages.
- Implement global exception handling via `GlobalExceptionHandler`.
- Add validation annotations in `TransactionDto`.
- Replace `double` type with `BigDecimal` for `quantity` in `TransactionDto`, `TransactionEntity`, `InventoryEntity`, and schema migration.
- Inject transaction validation into `TransactionServiceImpl`.

+++ src/main/java/com/comsystem/model/entity/InventoryEntity.java
@@ -29,7 +29,7 @@ public class InventoryEntity {
 
     private BigDecimal unitPrice;
 
-    private double quantity;
+    private BigDecimal quantity;
 
     private String unit;
 }
